chore(component): move tabs-pane to tabs/pane
