+++ src/components/tabs/pane/index.ts
