+++ src/index.ts
@@ -41,7 +41,7 @@ export { default as AtSegmentedControl } from './components/segmented-control'
 export { default as AtSwitch } from './components/switch'
 export { default as AtTabBar } from './components/tab-bar'
 export { default as AtTabs } from './components/tabs'
-export { default as AtTabsPane } from './components/tabs-pane'
+export { default as AtTabsPane } from './components/tabs/pane'
 export { default as AtTag } from './components/tag'
 export { default as AtTextarea } from './components/textarea'
 export { default as AtTimeline } from './components/timeline'
